// database/prisma-thinkthank/generated/zod/inputTypeSchemas/TokenOrderByWithAggregationInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';
import { SortOrderInputSchema } from './SortOrderInputSchema';
import { TokenCountOrderByAggregateInputSchema } from './TokenCountOrderByAggregateInputSchema';
import { TokenMaxOrderByAggregateInputSchema } from './TokenMaxOrderByAggregateInputSchema';
import { TokenMinOrderByAggregateInputSchema } from './TokenMinOrderByAggregateInputSchema';

export const TokenOrderByWithAggregationInputSchema: z.ZodType<Prisma.TokenOrderByWithAggregationInput> = z.object({
  id: z.lazy(() => SortOrderSchema).optional(),
  name: z.lazy(() => SortOrderSchema).optional(),
  hashedKey: z.lazy(() => SortOrderSchema).optional(),
  partialKey: z.lazy(() => SortOrderSchema).optional(),
  expires: z.union([ z.lazy(() => SortOrderSchema),z.lazy(() => SortOrderInputSchema) ]).optional(),
  lastUsed: z.union([ z.lazy(() => SortOrderSchema),z.lazy(() => SortOrderInputSchema) ]).optional(),
  createdAt: z.lazy(() => SortOrderSchema).optional(),
  updatedAt: z.lazy(() => SortOrderSchema).optional(),
  userId: z.lazy(() => SortOrderSchema).optional(),
  _count: z.lazy(() => TokenCountOrderByAggregateInputSchema).optional(),
  _max: z.lazy(() => TokenMaxOrderByAggregateInputSchema).optional(),
  _min: z.lazy(() => TokenMinOrderByAggregateInputSchema).optional()
}).strict();

export default TokenOrderByWithAggregationInputSchema;
